# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "title": "go-boilerplate API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/v1/purchase/list/{accountId}": {
            "get": {
                "description": "Get GetWalletListRes List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GetWalletListRes"
                ],
                "summary": "Get GetWalletListRes List",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Account ID",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-boilerplate_app_domain_wallet_dto.GetWalletListRes"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "go-boilerplate_app_domain_wallet_dto.GetWalletListRes": {
            "type": "object",
            "properties": {
                "bill_status": {
                    "type": "integer"
                },
                "currency_type": {
                    "description": "tb_bill",
                    "type": "integer"
                },
                "device_agent": {
                    "type": "string"
                },
                "download_count": {
                    "type": "integer"
                },
                "end_ts": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "renew": {
                    "type": "integer"
                },
                "start_ts": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "type": {
                    "type": "integer"
                },
                "wallet_status": {
                    "type": "integer"
                },
                "wallet_type": {
                    "description": "tb_wallet_product",
                    "type": "integer"
                }
            }
        }
    }
}